458 add more fall off to unique organization entity if there is only one

+++ src/Controller/OperatorController.php
@@ -2,6 +2,8 @@
 
 namespace App\Controller;
 
+use \Psr\Log\LoggerInterface;
+
 use Doctrine\DBAL\Exception\UniqueConstraintViolationException;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\ManagerRegistry;
@@ -9,11 +11,15 @@ use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Serializer\SerializerInterface;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
+
+use Symfony\Contracts\Cache\CacheInterface;
 
 use App\Form\OperatorType;
 use App\Form\TeamType;
@@ -25,9 +31,97 @@ use App\Entity\Trainer;
 use App\Entity\Team;
 use App\Entity\Uap;
 
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
-class OperatorController extends FrontController
+
+use App\Repository\UploadRepository;
+use App\Repository\ValidationRepository;
+use App\Repository\UapRepository;
+use App\Repository\TeamRepository;
+use App\Repository\OperatorRepository;
+use App\Repository\TrainingRecordRepository;
+use App\Repository\TrainerRepository;
+
+use App\Service\EntityDeletionService;
+use App\Service\EntityFetchingService;
+use App\Service\TrainingRecordService;
+use App\Service\PdfGeneratorService;
+
+
+class OperatorController extends AbstractController
 {
+
+    private $em;
+    private $request;
+    private $logger;
+    private $authChecker;
+    private $cache;
+
+    // Repository methods
+    private $validationRepository;
+    private $uploadRepository;
+    private $uapRepository;
+    private $teamRepository;
+    private $operatorRepository;
+    private $trainingRecordRepository;
+    private $trainerRepository;
+
+
+    // Services methods
+    private $entitydeletionService;
+    private $trainingRecordService;
+    private $pdfGeneratorService;
+    private $entityFetchingService;
+
+
+
+    public function __construct(
+
+        EntityManagerInterface          $em,
+        LoggerInterface                 $logger,
+        AuthorizationCheckerInterface   $authChecker,
+        RequestStack                    $requestStack,
+        CacheInterface                  $cache,
+
+        // Repository classes
+        ValidationRepository            $validationRepository,
+        UploadRepository                $uploadRepository,
+        UapRepository                   $uapRepository,
+        TeamRepository                  $teamRepository,
+        OperatorRepository              $operatorRepository,
+        TrainingRecordRepository        $trainingRecordRepository,
+        TrainerRepository               $trainerRepository,
+
+        // Services classes
+        EntityDeletionService           $entitydeletionService,
+        TrainingRecordService           $trainingRecordService,
+        PdfGeneratorService             $pdfGeneratorService,
+        EntityFetchingService           $entityFetchingService,
+
+    ) {
+        // $this->cache                        = $cache;
+
+        $this->em                           = $em;
+        $this->logger                       = $logger;
+        $this->authChecker                  = $authChecker;
+        $this->request                      = $requestStack->getCurrentRequest();
+        $this->cache                        = $cache;
+
+        // Variables related to the repositories
+        $this->validationRepository         = $validationRepository;
+        $this->uploadRepository             = $uploadRepository;
+        $this->uapRepository                = $uapRepository;
+        $this->teamRepository               = $teamRepository;
+        $this->operatorRepository           = $operatorRepository;
+        $this->trainingRecordRepository     = $trainingRecordRepository;
+        $this->trainerRepository            = $trainerRepository;
+
+        // Variables related to the services
+        $this->entitydeletionService        = $entitydeletionService;
+        $this->trainingRecordService        = $trainingRecordService;
+        $this->pdfGeneratorService          = $pdfGeneratorService;
+        $this->entityFetchingService        = $entityFetchingService;
+    }
 
     private function operatorEntitySearch(Request $request): array
     {
@@ -105,7 +199,6 @@ class OperatorController extends FrontController
                 ])->createView();
             }
         }
-        $flashes = $request->getSession()->getFlashBag()->all();
         // $this->logger->info('message in flashbag', $flashes);
 
         return $this->render('services/operators/operators_admin.html.twig', [
@@ -116,7 +209,7 @@ class OperatorController extends FrontController
 
 
 
-    
+
     private function processNewOperator(Operator $newOperator, $form, Request $request)
     {
 
@@ -136,6 +229,7 @@ class OperatorController extends FrontController
         $operator = $form->getData();
         $this->em->persist($operator);
         $this->em->flush();
+        $this->cache->delete('operators_list');
     }
 
 
@@ -146,7 +240,7 @@ class OperatorController extends FrontController
     #[Route('/operator/edit/{id}', name: 'app_operator_edit')]
     public function editOperatorAction(Request $request, Operator $operator): Response
     {
-        // $this->logger->info('Full request editOperatorAction', $request->request->all());
+        $this->logger->info('Full request editOperatorAction', $request->request->all());
 
         $form = $this->createForm(OperatorType::class, $operator, [
             'operator_id' => $operator->getId(),
@@ -156,19 +250,27 @@ class OperatorController extends FrontController
 
         if ($form->isSubmitted() && $form->isValid()) {
             $trainerBool = $form->get('isTrainer')->getData();
+            $this->logger->info('isTrainer form key value', [$trainerBool]);
+
             if ($trainerBool == true) {
                 if ($operator->getTrainer() == null) {
                     $trainer = new Trainer();
                     $trainer->setOperator($operator);
-                    $this->em->persist($trainer);
                     $operator->setTrainer($trainer);
+                } else {
+                    $trainer = $operator->getTrainer();
+                    $trainer->setDemoted(false);
                 }
+                $this->em->persist($trainer);
             } else {
                 $trainer = $operator->getTrainer();
-                $operator->setTrainer(null);
-                if ($trainer != null) {
+                if ($trainer != null && $trainer->getTrainingRecords() != null) {
+                    $trainer->setDemoted(true);
+                } else {
+                    $operator->setIsTrainer(false);
                     $this->em->remove($trainer);
                 }
+                $this->em->persist($trainer);
             }
             if ($operator->getTobedeleted() != null) {
                 $operator->setTobedeleted(null);
@@ -179,6 +281,8 @@ class OperatorController extends FrontController
             try {
                 $this->em->persist($operator);
                 $this->em->flush();
+                $this->cache->delete('operators_list');
+
                 // $this->logger->info('editOperatorAction, operateur bien modifié', [$operator]);
                 $this->addFlash('success', 'L\'opérateur a bien été modifié');
 
@@ -194,7 +298,7 @@ class OperatorController extends FrontController
                         'operator_id' => $operator->getId(),
                     ])->createView();
                 }
-                
+
                 // Return the updated frame content
                 return $this->render('services/operators/admin_component/_adminListOperator.html.twig', [
                     'operatorForms' => $operatorForms,
@@ -299,6 +403,9 @@ class OperatorController extends FrontController
         return $this->render('services/operators/operatorTraining.html.twig', [
             'trainingRecords'   => $trainingRecords,
             'upload'            => $upload,
+            'teams'             => $this->entityFetchingService->getTeams(),
+            'uaps'              => $this->entityFetchingService->getUaps(),
+            'operators'         => $this->entityFetchingService->getOperators(),
         ]);
     }
 
@@ -391,6 +498,7 @@ class OperatorController extends FrontController
 
         $this->em->persist($operator);
         $this->em->flush();
+        $this->cache->delete('operators_list');
         $this->addFlash('success', 'L\'opérateur a bien été ajouté');
         return $this->redirectToRoute('app_render_training_records', [
             'uploadId' => $uploadId,
@@ -562,6 +670,8 @@ class OperatorController extends FrontController
 
                 // Flush changes for each operator
                 $this->em->flush();
+                $this->cache->delete('operators_list');
+
             }
         }
 
@@ -882,7 +992,7 @@ class OperatorController extends FrontController
     #[Route('/operator/detail/{operatorId}', name: 'app_operator_detail')]
     public function printOpeDetail(int $operatorId)
     {
-        $operator = $this->cacheService->getEntityById('operator', $operatorId);
+        $operator = $this->operatorRepository->findOneBy(['id', $operatorId]);
         // $this->logger->info('operator', [$operator]);
 
         // $pdfContent = $this->pdfGeneratorService->generateOperatorPdf($operator);
@@ -979,6 +1089,7 @@ class OperatorController extends FrontController
                     try {
                         $em->persist($operator);
                         $em->flush();
+                        
                         break; // Exit loop if successful
                     } catch (UniqueConstraintViolationException $e) {
                         // Modify the violating field and retry
@@ -1008,6 +1119,7 @@ class OperatorController extends FrontController
             if (!$em->isOpen()) {
                 $em = $doctrine->resetManager();
             }
+            $this->cache->delete('operators_list');
 
             // Re-throw the exception for further handling
             throw $e;
@@ -1074,8 +1186,10 @@ class OperatorController extends FrontController
     #[Route('/operator/operator_management_base_page', name: 'app_operator_team_or_uap_management')]
     public function operatorManagement(Request $request): Response
     {
+        $teams = $this->teamRepository->findAll();
+        $uaps = $this->uapRepository->findAll();
 
-        if (count($this->teamRepository->findAll()) == 0 || count($this->uapRepository->findAll()) == 0) {
+        if (count($teams) == 0 || count($uaps) == 0) {
             $team = new Team();
             $uap = new Uap();
             $team->setName('INDEFINI');
@@ -1129,6 +1243,8 @@ class OperatorController extends FrontController
             }
         } else if ($request->getMethod() == 'GET') {
             return $this->render('services/operators/team_uap_operator_management.html.twig', [
+                'teams' => $teams,
+                'uaps' => $uaps,
                 'teamForm' => $teamForm->createView(),
                 'uapForm' => $uapForm->createView()
             ]);
